pyproject.toml
[project]
name = "faultsense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-mistralai>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pypdf>=6.9.1",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
